refactor: move Toast to ArchivedComponents
