+++ src/components/ArchivedComponents/Toast/Toast.test.tsx
@@ -1,11 +1,10 @@
 import userEvent from "@testing-library/user-event";
 
-import ToastProvider from "../../context/Toast/ToastProvider";
-import renderWithTheme from "../../__tests__/__helpers__/renderWithTheme";
-import waitForNextTick from "../../__tests__/__helpers__/waitForNextTick";
-
 import Toast from "./Toast";
 
+import ToastProvider from "@/context/Toast/ToastProvider";
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+import waitForNextTick from "@/__tests__/__helpers__/waitForNextTick";
 import CopyLinkButton from "@/components/SharedComponents/Button/CopyLinkButton";
 
 describe("toast notification", () => {
